Fixed coderabbit issue
